Move clues into respective scene packages
