+++ src/main/java/clue/firstscene/FatherInsurance.java
@@ -1,4 +1,6 @@
-package clue;
+package clue.firstscene;
+
+import clue.Clue;
 
 public class FatherInsurance extends Clue {
