+++ src/main/java/clue/firstscene/FatherMap.java
@@ -1,4 +1,6 @@
-package clue;
+package clue.firstscene;
+
+import clue.Clue;
 
 public class FatherMap extends Clue {
