+++ src/test/java/suspect/SuspectListTest.java
@@ -5,9 +5,9 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import clue.Clue;
-import clue.FatherInsurance;
-import clue.FatherMap;
-import clue.FatherTextMessage;
+import clue.firstscene.FatherInsurance;
+import clue.firstscene.FatherMap;
+import clue.firstscene.FatherTextMessage;
 import org.junit.jupiter.api.Test;
 import seedu.duke.Ui;
